+++ ramutils/cli/__init__.py
@@ -4,10 +4,7 @@ import os.path
 
 from ramutils import __version__
 from ramutils.constants import EXPERIMENTS
-
-
-class ValidationError(Exception):
-    """Raised when command-line arguments are invalid."""
+from ramutils.exc import ValidationError  # so that other imports don't need to be updated
 
 
 class RamArgumentParser(ArgumentParser):